Implemented WW Timer Library with Play, Stop, and Pause APIs via WebSocket (#1)
* Feat: created WW Timer

* Other: Added eslint rules

* Other: Eslint 규칙 추가

* Other: Added husky

* BREAKING CHANGES: changed 'play' method name to 'start'

* Docs: API 변경 사항 반영

* Other: updated for npm deploy

* Other: keywords 정의

* Refactor: Rename play method to start in WWTimer class

This commit renames the `play` method to `start` in the WWTimer class to better reflect the action being performed by the method. The change aims to improve code readability and make the method's purpose more intuitive, especially for developers familiar with common naming conventions for starting processes or activities. All references to the `play` method have been updated to `start`, ensuring consistency throughout the codebase. No functionality has been altered by this renaming.

* Add: rollup configuration for ES module bundling

Added a rollup.config.js file to configure the bundling process for the WWTimer project. This configuration specifies 'src/index.js' as the entry point and outputs the bundled file in ES module format to 'dist/ww-timer.js'. This setup ensures the project is bundled efficiently for use with modern JavaScript module systems.

* Add: workflow for deploying package to NPM with version confirmation

This commit introduces a new GitHub Actions workflow to automate the process of deploying our package to NPM. It requires a manual trigger through the GitHub UI, where the user is prompted to confirm the version they wish to deploy. This ensures that deployments are intentional and verified. The workflow includes steps for checking out the code, setting up Node.js, validating the tag, installing dependencies, linting, building, and finally deploying the package to NPM.

* Update: build script and add Rollup as a dev dependency

- Simplified the "build" script in package.json to use the rollup configuration file by changing it to "rollup -c". This change leverages the rollup.config.js file for the build process, making the build script cleaner and more maintainable.
- Added Rollup as a development dependency to ensure a consistent build process across different environments. This addition aids in maintaining the build setup within the project's ecosystem without requiring

* Add: registry-url and NODE_AUTH_TOKEN for NPM deployment workflow

Enhanced the NPM deployment GitHub Actions workflow by specifying the `registry-url` for the `actions/setup-node` step and including the `NODE_AUTH_TOKEN` environment variable in the deploy step. These changes ensure that the workflow has the necessary configurations to authenticate and publish the package to the NPM registry securely. The `registry-url` is set to "https://registry.npmjs.org", aligning the deployment process with the official NPM registry

* Remove: deploy script from package.json

The "deploy" script has been removed from the `package.json` to streamline the project's scripts. This change reflects a shift towards utilizing external CI/CD pipelines or manual commands for deployment processes, adhering to a cleaner and more focused `package.json` structure. Adjustments were made to ensure the project's build and development lifecycle scripts remain concise and relevant to common development workflows.

* Update: deployment command to use npm directly in GitHub Actions

Changed the deployment command in the 'Deploy package on NPM' GitHub Actions workflow from `yarn publish` to `npm publish`. This adjustment ensures direct utilization of npm for the publishing process, leveraging the `NODE_AUTH_TOKEN` environment variable for authentication. This modification aligns the deployment step more closely with npm

* Add: Added LISENCE File