Update .npmignore to Correctly Exclude .husky Directory (#6)
* Fix: Exclude examples directory from version control

To keep the repository clean and focused on the source code, this commit updates .gitignore and .npmignore to exclude the examples directory. The examples directory contains real-world usage examples of the library, which are not necessary for the package functionality but useful for developers. By excluding these files from version control, we minimize the package size and streamline the development process.

* Update: API Reference in README to Reflect Revised API Structure

This commit updates the API Reference section of the README.md file to accurately reflect the changes made to the WWTimer API structure. It clarifies the usage of the start(), pause(), and destroy() methods according to the new implementation where the callback function and interval are specified at the instance creation. This change ensures that the documentation is in sync with the current API design, making it easier for users to understand how to implement and use the WWTimer in their projects.

* v1.0.2

* Update: Optimize package size by refining .npmignore

Updated the .npmignore file to exclude additional files and directories such as .github/, .eslintrc.js, husky/, src/, and rollup.config.mjs along with banner.jpg. This change aims to further reduce the size of the build package, ensuring a more efficient distribution and installation process.

* v1.0.3

* Update: .npmignore to Correctly Exclude .husky Directory

In an effort to ensure our project's npm package remains clean and focused, this commit updates the .npmignore file to correctly exclude the `.husky` directory by adding the missing dot prefix. Previously, the `husky/` entry without the leading dot may have inadvertently allowed Husky's configuration files to be included in the packaged distribution. By specifying `.husky/`, we prevent these development-specific files from being unnecessarily published, aligning with our goal to streamline the package contents and maintain a clear separation between development tools and the published code.